Fix bug 1057286: Split file attachments from wiki into a separate application.
