+++ kuma/attachments/templates/attachments/attachment_detail.html
