+++ kuma/attachments/templates/attachments/includes/attachment_list.html
@@ -37,7 +37,7 @@
 
 
   {% if show_attach_button and attachment_form %}
-  <form method="post" action="{{ url('wiki.new_attachment') }}" id="page-attachments-form" enctype="multipart/form-data">
+  <form method="post" action="{{ url('attachments.new_attachment') }}" id="page-attachments-form" enctype="multipart/form-data">
     {{ csrf() }}
     <table cellpadding="0" cellspacing="0" id="page-attachments-new-table" class="hidden">
       <thead>
