+++ kuma/attachments/templates/attachments/includes/attachment_upload_results.html
