+++ kuma/wiki/feeds.py
@@ -16,7 +16,7 @@ from django.utils.translation import ugettext as _
 from sumo.urlresolvers import reverse
 from kuma.users.models import UserProfile
 from .helpers import diff_table, tag_diff_table, compare_url, colorize_diff
-from .models import Document, Revision, AttachmentRevision
+from .models import Document, Revision
 
 
 MAX_FEED_ITEMS = getattr(settings, 'MAX_FEED_ITEMS', 500)
@@ -359,39 +359,6 @@ class RevisionsFeed(DocumentsFeed):
         return self.request.build_absolute_uri(
             reverse('kuma.wiki.views.document', locale=item.document.locale,
                     args=(item.document.slug,)))
-
-    def item_pubdate(self, item):
-        return item.created
-
-    def item_author_name(self, item):
-        return '%s' % item.creator
-
-    def item_author_link(self, item):
-        return self.request.build_absolute_uri(item.creator.get_absolute_url())
-
-    def item_categories(self, item):
-        return []
-
-
-class AttachmentsFeed(DocumentsFeed):
-    title = _("MDN recent file changes")
-    subtitle = _("Recent revisions to MDN file attachments")
-
-    def items(self):
-        return AttachmentRevision.objects.order_by('-created')[:50]
-
-    def item_title(self, item):
-        return item.title
-
-    def item_description(self, item):
-        previous = item.get_previous()
-        if previous is None:
-            return '<p>Created by: %s</p>' % item.creator.username
-        return "<p>Edited by %s: %s" % (item.creator.username, item.comment)
-
-    def item_link(self, item):
-        return self.request.build_absolute_uri(
-            item.attachment.get_absolute_url())
 
     def item_pubdate(self, item):
         return item.created
