+++ kuma/wiki/templates/wiki/document.html
@@ -338,7 +338,7 @@
 
                 <!-- attachments list -->
                 {% if document.attachments.all().exists() %}
-                  {% include 'wiki/includes/attachment_list.html' %}
+                  {% include 'attachments/includes/attachment_list.html' %}
                 {% endif %}
 
                 <!-- contributors -->
