+++ kuma/wiki/templates/wiki/edit_document.html
@@ -192,7 +192,7 @@
       </section>
     </form>
 
-    {% include 'wiki/includes/attachment_list.html' %}
+    {% include 'attachments/includes/attachment_list.html' %}
 
     <!-- Add information about this document in JSON format for use in CKEditor -->
     <script type="text/javascript">
