+++ kuma/wiki/templates/wiki/new_document.html
@@ -84,7 +84,7 @@
       </section>
     </form>
 
-    {% include 'wiki/includes/attachment_list.html' %}
+    {% include 'attachments/includes/attachment_list.html' %}
 
     </article>
