+++ kuma/wiki/templates/wiki/translate.html
@@ -161,7 +161,7 @@
       </div>
     </form>
 
-  {% include 'wiki/includes/attachment_list.html' %}
+  {% include 'attachments/includes/attachment_list.html' %}
 
 {% endblock %}
